Use the panel "shadowtest" for best results -- a working SSAO implementation!!
The one catch: Because we don't have float textures with WEBGL, there is
horrid banding, an _incredibly_ narrow effective range, and terrrible
artifacts when the camera is at an angle. So basically, I believe this to be
working, but there's not much to look at.

Also, it still needs some blurring or downsampling or other postprocessing
to reduce the noisiness of it, but that's a secondary concern.